requirements.txt
fastapi
uvicorn
pydantic
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pandas
motor